Move Unserializer into Serialization namespace + add interface
