+++ src/Serialization/ValueUnserializer.php
@@ -1,0 +1,9 @@
+<?php
+namespace Prezly\Slate\Serialization;
+
+use Prezly\Slate\Model\Value;
+
+interface ValueUnserializer
+{
+    public function fromJson(string $json): Value;
+}
